requirements.txt
gradio>=4.0
gradio-client>=0.9
beautifulsoup4>=4.12
python-dotenv>=1.0
pytest>=8.0
websockets>=10.0
